DAY20
Given two strings s and t. Find the minimum number of operations that need to be performed on str1 to convert it to str2. The possible operations are:

Insert a character at any position of the string.
Remove any character from the string.
Replace any character from the string with any other character.